S20-579: refactor select component with hoc
